% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressAllMS.R
\name{compressAllMS}
\alias{compressAllMS}
\title{Compress a set of DMS files.}
\usage{
compressAllMS(
  origMsDir,
  compMsDir,
  ms_type = c("fticr", "esquire"),
  dataRepo = "../data/",
  compMode = c("targets", "grid", "grid+targets"),
  tgTable = NA,
  dmzTarget = 0.5,
  mzMin = 70,
  mzMax = 250,
  dmz = 0.001,
  test = FALSE
)
}
\arguments{
\item{origMsDir}{(string; mandatory) Path to repertory of DMS files
to be compressed}

\item{compMsDir}{(string; mandatory) Path to repertory of compressed files}

\item{ms_type}{(string; optional) Type of mass spectrum,
one of `esquire` or `fticr` (default).
Compression only for `fticr` MS.}

\item{dataRepo}{(string; optional) Path to the data files.}

\item{compMode}{(string; optional) Compression mode:
either `grid`, `targets` (default),
or their combination `grid+targets`.}

\item{tgTable}{(string; optional)  Path to targets file.}

\item{dmzTarget}{(numeric; optional) Half-width of m/z interval
centered on target reference m/z value.}

\item{mzMin}{(numeric; optional) Min m/z value of compression grid}

\item{mzMax}{(numeric; optional) Max m/z value of compression grid}

\item{dmz}{(numeric; optional) m/z step or compression grid}

\item{test}{(logical; optional) if TRUE, shodt run to check the
arguments validity and the compression efficiency}
}
\value{
Returns nothing. It is used for its side effects.
}
\description{
Compress a set of DMS files.
}
