% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressMS.R
\name{compressMS}
\alias{compressMS}
\title{Down-sample a FT-ICR MS by choosing points closest to a preset
regular grid or around a set of targets.}
\usage{
compressMS(
  file_in,
  file_out,
  ms_type = "fticr",
  mzMin = 70,
  mzMax = 250,
  dmz = 0.002,
  test = FALSE,
  compMode = "grid+targets",
  targets = NA,
  dmzTarget = 1
)
}
\arguments{
\item{dmz}{}
}
\value{
Nothing. Used for its side effects.
}
\description{
Down-sample a FT-ICR MS by choosing points closest to a preset
regular grid or around a set of targets.
}
