% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmsAnalysis.R
\name{dmsAnalysis}
\alias{dmsAnalysis}
\title{Analysis of DMS file}
\usage{
dmsAnalysis(
  ms_type = c("esquire", "fticr")[1],
  taskTable = NA,
  tgTable = NA,
  dataRepo = "../data/",
  figRepo = "../results/figs/",
  tabRepo = "../results/tables/",
  fit_dim = 1,
  filter_results = TRUE,
  userTag = paste0("fit_dim_", fit_dim),
  save_figures = TRUE,
  plot_maps = FALSE,
  fallback = TRUE,
  correct_overlap = FALSE,
  weighted_fit = FALSE,
  refine_CV0 = TRUE,
  debug = FALSE
)
}
\arguments{
\item{debug}{}
}
\value{

}
\description{
Analysis of DMS file
}
