% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit1D.R
\name{fit1D}
\alias{fit1D}
\title{1D peak fit along CV axis}
\usage{
fit1D(
  mz0,
  CV0,
  dmz,
  dCV,
  mz,
  CV,
  MS,
  fwhm_cv_nom,
  weighted = FALSE,
  refine_CV0 = FALSE,
  correct_overlap = FALSE
)
}
\arguments{
\item{correct_overlap}{}
}
\value{

}
\description{
1D peak fit along CV axis
}
