% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit1D_MS.R
\name{fit1D_MS}
\alias{fit1D_MS}
\title{1D peak fit along m/z axis}
\usage{
fit1D_MS(mz0, CV0, dmz, dCV, mz, CV, MS, fwhm_mz_nom, weighted = FALSE)
}
\arguments{
\item{weighted}{}
}
\value{

}
\description{
1D peak fit along m/z axis
}
