% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwm.R
\name{fwm}
\alias{fwm}
\title{Weighted mean and uncertainty by the Dersimonian-Laird method.}
\usage{
fwm(x, ux)
}
\arguments{
\item{x}{(vector) set of values to be averages.}

\item{ux}{(vector) uncertainties of values to be averaged.}
}
\value{
A list with elements:
\describe{
  \item{wm}{Weighted mean.}
  \item{uwm}{Uncertainty of the weighted mean.}
}
}
\description{
A dark uncertainty component is added for the statistical
compatibility of the averaged values.
}
