% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherResults.R
\name{gatherResults}
\alias{gatherResults}
\title{Gather results in data frame}
\usage{
gatherResults(Tasks, tabRepo, userTag)
}
\arguments{
\item{userTag}{}
}
\value{

}
\description{
Gather results in data frame
}
