% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMS.R
\name{getMS}
\alias{getMS}
\title{Get MS from file}
\usage{
getMS(file, ms_type = c("esquire", "fticr"))
}
\arguments{
\item{file}{(string; mandatory) Path to DMS file}

\item{ms_type}{(string; optional) Type of mass spectrum,
one of `esquire` (default) or `fticr`.}
}
\value{
List of three elements: `mz` = vector of m/z coordinates,
  `time` = vector of time coordinates, and `MS` = DMS matrix.
}
\description{
Get MS from file
}
