% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPars2D.R
\name{getPars2D}
\alias{getPars2D}
\title{Extract parameters from 2D Gaussian peak fit}
\usage{
getPars2D(res)
}
\arguments{
\item{res}{(nls-object) a nls fit result}
}
\value{
A list of best-fit parameters:
\describe{
  \item{v}{vector of Gaussian peak best-fit parameters}
  \item{u_v}{uncertainty on v elements}
  \item{mzopt}{peak center along m/z}
  \item{u_mz}{uncertainty on mzopt}
  \item{cvopt}{peak center along CV}
  \item{u_cv}{uncertainty on cvopt}
  \item{fwhm_mz}{peak FWHM along m/z}
  \item{u_fwhm_mz}{uncertainty on fwhm_mz}
  \item{fwhm_cv}{peak FWHM along CV}
  \item{u_fwhm_cv}{uncertainty on fwhm_cv}
  \item{area}{peak area}
  \item{u_area}{uncertainty on area}
}
}
\description{
Extract parameters from 2D Gaussian peak fit
}
