% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTag.R
\name{makeTag}
\alias{makeTag}
\title{Create a tag from DMS and CV filenames and a user tag.}
\usage{
makeTag(CVTable, msTable, userTag)
}
\arguments{
\item{CVTable}{(string) name of CV file.}

\item{msTable}{(string) name of ms file.}

\item{userTag}{(string) string to append to results filenames.}
}
\description{
Create a tag from DMS and CV filenames and a user tag.
}
