% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaps.R
\name{plotMaps}
\alias{plotMaps}
\title{Plot DMS fit maps}
\usage{
plotMaps(
  mz,
  CV,
  MS,
  mex = NA,
  leg = NA,
  tag = NA,
  mzlim = range(mz),
  CVlim = range(CV),
  zlim = c(0, max(MS)/2),
  logz = TRUE,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Plot DMS fit maps
}
