% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz}
\alias{trapz}
\title{Trapezoidal integration}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{(vector) x values}

\item{y}{(vector) y = f(x) values}
}
\value{
The trapezpoidal integration value.
}
\description{
Trapezoidal integration
}
