% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2D.R
\name{fit2D}
\alias{fit2D}
\title{2D peak fit}
\usage{
fit2D(
  mz0,
  CV0,
  dmz,
  dCV,
  mz,
  CV,
  MS,
  fwhm_mz_nom,
  fwhm_cv_nom,
  weighted = FALSE,
  refine_CV0 = FALSE,
  correct_overlap = FALSE
)
}
\arguments{
\item{correct_overlap}{}
}
\value{

}
\description{
2D peak fit
}
