% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherResults.R
\name{gatherResults}
\alias{gatherResults}
\title{Gather all results generated by an analysis into a single dataframe.}
\usage{
gatherResults(Tasks, tabRepo, userTag)
}
\arguments{
\item{Tasks}{(dataframe)}

\item{tabRepo}{(string) Path to output tables.}

\item{userTag}{(string) String to append to results filenames.}
}
\value{
A dataframe with same columns as "xxx_results.csv".
}
\description{
Gather all results generated by an analysis into a single dataframe.
}
