% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakSpecs.R
\name{getPeakSpecs}
\alias{getPeakSpecs}
\title{Get apparatus-based specifications for peaks}
\usage{
getPeakSpecs(ms_type)
}
\arguments{
\item{ms_type}{}
}
\value{

}
\description{
Get apparatus-based specifications for peaks
}
