% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaps.R
\name{plotMaps}
\alias{plotMaps}
\title{Plot DMS fit maps}
\usage{
plotMaps(
  mz,
  CV,
  MS,
  mex = NA,
  leg = NA,
  tag = NA,
  mzlim = range(mz),
  CVlim = range(CV),
  zlim = c(0, max(MS)/2),
  logz = TRUE,
  gPars
)
}
\arguments{
\item{mz}{tbd}

\item{CV}{tbd}

\item{MS}{tbd}

\item{mex}{tbd}

\item{leg}{tbd}

\item{tag}{tbd}

\item{mzlim}{tbd}

\item{CVlim}{tbd}

\item{zlim}{tbd}

\item{logz}{tbd}

\item{gPars}{tbd}
}
\value{
tbd
}
\description{
Plot DMS fit maps
}
