% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeak.R
\name{plotPeak}
\alias{plotPeak}
\title{Plot peak fit results}
\usage{
plotPeak(
  mz,
  CV,
  MS,
  fitOut,
  mex = NA,
  leg = NA,
  val = NA,
  tag = NA,
  type = "CV",
  CV0 = NA,
  gPars
)
}
\arguments{
\item{gPars}{}
}
\value{

}
\description{
Plot peak fit results
}
